get a users holds in the PhysicalItemDetails component so we can tell the user if it is them that has an item on hold
